requirements.txt
apache-airflow-providers-apache-spark==4.9.0
apache-airflow-providers-amazon
apache-airflow-providers-google
pyspark==3.5.2
pymysql
pandas
mlflow